modified plugin structure
